Added EntityManager, fixed slimes and stuff

+++ src/main/java/org/egordorichev/lasttry/entity/enemy/Boss.java
@@ -1,14 +1,12 @@
 package org.egordorichev.lasttry.entity.enemy;
 
-import org.egordorichev.lasttry.entity.Enemy;
-
 public class Boss extends Enemy {
 	protected String name;
 
 	protected Phase[] phases;
 	protected Phase currentPhase;
 
-	public Boss(int id, String name, int maxHp) {
+	public Boss(short id, String name, int maxHp) {
 		super(id);
 		this.name = name;
 	}
